Migrate to only deploying the cloud environment (#10)
* Convert to client side only

Drop all the work and deployment stuff related to installation on kube-ansible
since we've ported everything over to OpenShift Ansible. Migrates to simply
deploying the cloud monitoring side of things and having it connect to the remote
central QDR.

* Tweaks and changes to barometer deployment

* Drop the contrib/ directory

We're not longer doing work against kube-ansible. You should see the work done
in redhat-nfvpe/openshift-ansible for deployment now.

* Update README for deployment