Update client side integration
Update the client side integration to use the latest image versions and to
deploy the configuration in a more consistent manner that matches how
OpenStack deploys the configuration for kolla-built containers.

Validation was performed against a minishift deployment running
Service Assurance. Traffic was from a locally running instance of
kolla-build collectd as shipped upstream in RDO. Deployment of
qdrouterd is still using the shipped image from InterconnectedCloud
due to a known issue with use of edge connections that is not yet
shipped in the upstream kolla-based container images.